Promote discrete constraints to stable API
