+++ gtsam/discrete/AllDiff.cpp
@@ -6,8 +6,8 @@
  */
 
 #include <gtsam/base/Testable.h>
-#include <gtsam_unstable/discrete/AllDiff.h>
-#include <gtsam_unstable/discrete/Domain.h>
+#include <gtsam/discrete/AllDiff.h>
+#include <gtsam/discrete/Domain.h>
 
 #include <optional>
