+++ gtsam_unstable/discrete/CSP.h
@@ -8,8 +8,9 @@
 #pragma once
 
 #include <gtsam/discrete/DiscreteFactorGraph.h>
-#include <gtsam_unstable/discrete/AllDiff.h>
-#include <gtsam_unstable/discrete/SingleValue.h>
+#include <gtsam/discrete/AllDiff.h>
+#include <gtsam/discrete/SingleValue.h>
+#include <gtsam_unstable/dllexport.h>
 
 namespace gtsam {
